server/requirements.txt
fastapi>=0.112
uvicorn[standard]>=0.30
openai>=1.0
pyyaml>=6.0
pydantic>=2.0
garminconnect
